requirements.txt
aiohttp==3.9.3
influxdb_client==1.41.0
python-dotenv==1.0.1 
